add icons through font-awesome and bootstrap social
This commit adds the css from font awesome and bootstrap social, then
implements the icons in the nav-bar as well as in the footer. The social
icons in the footer are a little bit big, especially when compared to the
mail icon. This needs fixing. other than that, though, all else is well.